move CamControl to SoftcamSetup
